chore: organize code in view dir
