+++ src/views/product-downloads-view/boundary/components/desktop-client-callout/desktop-client-callout.module.css
@@ -4,8 +4,3 @@
 	color: var(--token-color-foreground-strong);
 	margin-bottom: 16px;
 }
-
-.devPre {
-	border: 1px solid magenta;
-	white-space: pre-wrap;
-}
