+++ src/views/product-downloads-view/boundary/components/desktop-client-callout/types.ts
@@ -1,0 +1,13 @@
+import { OperatingSystem } from 'lib/fetch-release-data'
+
+export interface ReleaseBuild {
+	os: OperatingSystem
+	url: string
+	filename: string
+	arch: string
+}
+
+export interface DesktopClientProps {
+	latestVersion: string
+	builds: ReleaseBuild[]
+}
