+++ src/views/product-downloads-view/boundary/components/index.tsx
@@ -1,1 +1,3 @@
 export * from './icon-linux-color-16'
+export * from './desktop-client-callout'
+export * from './operating-system-icons'
